requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-filter==22.1
drf-spectacular[sidecar]==0.22.1
gitpython==3.1.27
# dj-rest-auth  # https://github.com/iMerica/dj-rest-auth
# djangorestframework-gis  # https://github.com/openwisp/django-rest-framework-gis